WiP: testing non-regression on arm32

+++ so3/include/syscall.h
@@ -91,7 +91,7 @@
 #include <errno.h>
 #include <types.h>
 
-int syscall_handle(uint32_t, uint32_t, uint32_t, uint32_t);
+long syscall_handle(unsigned long, unsigned long, unsigned long, unsigned long);
 
 void set_errno(uint32_t val);
 #endif /* __ASSEMBLY__ */
